+++ MANIFEST.in
@@ -1,2 +1,2 @@
-recursive-include popeui/ *
-recursive-include popegen/ *
+recursive-include engel/ *
+recursive-include eng/ *
