+++ docs/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# PopeUI documentation build configuration file, created by
+# Engel documentation build configuration file, created by
 # sphinx-quickstart on Sun Jul 17 00:03:45 2016.
 #
 # This file is execfile()d with the current directory set to its
@@ -53,7 +53,7 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'PopeUI'
+project = u'Engel'
 copyright = u'2016, William Dussault'
 author = u'William Dussault'
 
@@ -209,7 +209,7 @@ html_static_path = ['_static']
 #html_search_scorer = 'scorer.js'
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'PopeUIdoc'
+htmlhelp_basename = 'Engeldoc'
 
 # -- Options for LaTeX output ---------------------------------------------
 
@@ -231,7 +231,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-  (master_doc, 'PopeUI.tex', u'PopeUI Documentation',
+  (master_doc, 'Engel.tex', u'Engel Documentation',
    u'William Dussault', 'manual'),
 ]
 
@@ -261,7 +261,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'popeui', u'PopeUI Documentation',
+    (master_doc, 'engel', u'Engel Documentation',
      [author], 1)
 ]
 
@@ -275,8 +275,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-  (master_doc, 'PopeUI', u'PopeUI Documentation',
-   author, 'PopeUI', 'One line description of project.',
+  (master_doc, 'Engel', u'Engel Documentation',
+   author, 'Engel', 'One line description of project.',
    'Miscellaneous'),
 ]
