+++ docs/howto/engel.rst
@@ -1,0 +1,12 @@
+Quickstart Guide
+================
+
+This section aims to explain the basics of working with Engel. 
+It focuses primarily on application development, but will also touch the subject of extending Engel adding your own widgets and libraries.
+
+This guide assumes that you already have Engel installed and working in your environment. If that is not the case, you should check out the :doc:`../install` section.
+
+.. toctree::
+
+  engel.application
+  engel.widget
