+++ docs/howto/engel.widget.rst
@@ -1,0 +1,13 @@
+Extending Engel With Widgets
+=============================
+
+This section will give you an overview of how you can extend Engel by creating your own widgets. It assumes
+you already have a rudimentary comprehension of how to make a Engel application.
+
+We will be making a ``LoginForm`` widget, to be used in later projects.
+
+
+Generating the project files
+----------------------------
+
+Like in the previous tutorial, start by generating your project by using ``eng``, Engel's built-in code-generator.
