--- docs/howto/popeui.rst
@@ -1,12 +1,0 @@
-Quickstart Guide
-================
-
-This section aims to explain the basics of working with PopeUI. 
-It focuses primarily on application development, but will also touch the subject of extending PopeUI by adding your own widgets and libraries.
-
-This guide assumes that you already have PopeUI installed and working in your environment. If that is not the case, you should check out the :doc:`../install` section.
-
-.. toctree::
-
-  popeui.application
-  popeui.widget
