--- docs/howto/popeui.widget.rst
@@ -1,13 +1,0 @@
-Extending PopeUI With Widgets
-=============================
-
-This section will give you an overview of how you can extend PopeUI by creating your own widgets. It assumes
-you already have a rudimentary comprehension of how to make a PopeUI application.
-
-We will be making a ``LoginForm`` widget, to be used in later projects.
-
-
-Generating the project files
-----------------------------
-
-Like in the previous tutorial, start by generating your project by using ``bishop``, PopeUI's build-in code-generator.
