+++ docs/index.rst
@@ -1,9 +1,9 @@
-.. PopeUI documentation master file, created by
+.. Engel documentation master file, created by
    sphinx-quickstart on Sun Jul 17 00:03:45 2016.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to PopeUI's documentation!
+Welcome to Engel's documentation!
 ==================================
 
 Documentation
@@ -13,6 +13,6 @@ Documentation
    :maxdepth: 2
 
    install
-   howto/popeui
+   howto/engel
    support
-   code_overview/popeui
+   code_overview/engel
