+++ docs/support.rst
@@ -1,4 +1,4 @@
 Support
 =======
 
-Since this project is still in its infancy, the best way to get help with the project is by opening an issue on `Github <https://github.com/Dalloriam/popeui/issues>`_. We usually keep our issues updated, and you should get an answer from us pretty quickly.
+Since this project is still in its infancy, the best way to get help with the project is by opening an issue on `Github <https://github.com/Dalloriam/engel/issues>`_. We usually keep our issues updated, and you should get an answer from us pretty quickly.
