+++ eng/templates/app.template
@@ -1,4 +1,4 @@
-from popeui import Application
+from engel import Application
 
 from views.home import HomeView
 from services.home import HomeService
