+++ eng/templates/view.template
@@ -1,4 +1,4 @@
-from popeui import View
+from engel import View
 
 
 class {0}(View):
